pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dualmation"
version = "0.1.0"
description = "Hybrid LLM-Diffusion Framework for Educational Animation Generation with RL Feedback"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "nabin2004"},
]
keywords = ["animation", "manim", "llm", "diffusion", "reinforcement-learning", "educational"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
]

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "diffusers>=0.25.0",
    "accelerate>=0.25.0",
    "safetensors>=0.4.0",
    "einops>=0.7.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "trl>=0.7.0",
    "manim>=0.18.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
]

[project.urls]
Homepage = "https://github.com/nabin2004/dualmation"
Repository = "https://github.com/nabin2004/dualmation"
Issues = "https://github.com/nabin2004/dualmation/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 100
